requirements.txt
docopt==0.6.2
scipy==0.19.0
scikit-learn==0.18.1
matplotlib==2.0.0
tensorflow-gpu==1.0.1
tflearn==0.3
joblib==0.11